git: prevent crash during git clone
Summary:
If the git clone failed, it would throw an exception and try to clean
up. If there was no destination directory provided, the cleanup code would call
shutil(None) which would crash.

Additionally, if the directory already existed, an exception would be thrown and
the shutil would try to delete the existing directory. Which would've been data
loss. Luckily the earlier bug mentioned above prevented that.

Fix both of these by moving the destpath calculation and exists check to be
outside the try/catch.

Reviewed By: MichaelCuevas

Differential Revision: D41296351

fbshipit-source-id: dae00a8e09fa72cae8fd31d53fda4d88fefa28bc