update various command aliases and synopses
Summary:
- move silly aliases to the legacy list
- simplify some synopses

I left particular flags in the synopsis where I felt the flag was fundamental to the command, or the command had very few flags. Other than that, it doesn't really help much to see a bunch of single letter flags since you have to refer to the help text to see what they do anyway. (Also, the synopses are hand maintained, so they get out of sync easily).

Reviewed By: DurhamG

Differential Revision: D41292832

fbshipit-source-id: e95e70d7d2641bb49ee25683b6ad614d5ef08ae4