Fix pushredirection logic when previous bookmark value doesn't map directly to small repo
Summary:
To fix the bug we added a test for in the previous diff, we need to be careful when converting the old bookmark value from the large repo to the small one.

Even though the new bookmark value always exists in the small repo, the previous one doesn't necessarily do. It might have been a large-only change. So when we map the old bookmark value back, we should actually get the last small commit that had exactly the same changes in the small repo.

Notice that, even before, the push worked, but the command failed. Now both succeed.

Differential Revision: D41270747

fbshipit-source-id: ee19b73ae414467734b655f351840e58cdd97560