Speed up response conversion
Summary:
This speeds up the response conversion by doing all the large->small conversions at the same time. They unnecessarily used to happen serially.

This is fine because at this point everything has been synced. Notice that the `backsync_latest` command before **cannot** be done in parallel with the conversions.

Reviewed By: RajivTS

Differential Revision: D41273248

fbshipit-source-id: ec622256f3a135439adeb49e1f0f67ca9cf18bdc