refactor ImportPriority
Summary:
This diff improves ImportPriority in a variety of ways.

It was originally intended that ImportPriority fit in a single 64-bit
integer. Replace the bitfield with a single uint64_t and manual
shifts. Also, add a static_assert to enforce the intent.

Use semantic constants to make it clear the relative import priorities
across EdenFS.

Add some unit tests.

Reviewed By: genevievehelsel

Differential Revision: D40602313

fbshipit-source-id: 3c99249ef2863396235ff1159b10b54d7478671d