dag: provide a way to override default segment size
Summary:
The default segment size 16 requires about 16 merges to demonstrate Level 1
segments, and about 256 for Level 2.

The current API to set segment size is tricky to use since the IdDag is kind
of managed by NameDag.  To demonstrate the concepts using simpler graphs. Let's
add an API to tweak the default segment size.

Reviewed By: muirdm

Differential Revision: D41229584

fbshipit-source-id: 1daae28db04a809051bee5dbfe78538f41166032