fix scrolling comparison view
Summary:
It seems in the transition to <Modal> as a wrapper around the comaprison view, I broke scrolling long comparisons. This is super annoying so we definitely need to fix it.

The fix here uses `display: contents`, which sort-of makes the wrapping `div`s` as if they weren't there, which allows the <Modal> styling to act like it's the direct parent of the comparison view contents, which allows the parent height to be correct so the overflow-y: scroll works as expected.

`display: contents` seems pretty ok to use these days: https://caniuse.com/?search=display%20contents

Reviewed By: bolinfest

Differential Revision: D41345453

fbshipit-source-id: 7b2bd7a7dba04e3fb6c667cb09e3de96aee4d899