Propagate RepositoryError to ServerToClientAPI
Summary:
We recently changed RepoInfo so it could also fail with a RepositoryError union type. This apparently did not work for vscode, where the Repository is set up ahead of the ISL webview being created. We need to support the requestRepoInfo message being passed from the client so we can answer requests with the most recent RepoError, if a Repository is not successfully created.

This means we need ServerToClientAPI to contain either a repo or a repository error, rather than ONLY a Repository or null. Previously we assumed an invariant of having the repository in order to handle messages. This is not really a good assumption— we should be able to handle SOME messages even if there's no repo. This will let us support requesting errors or in the future handling requests to change the CWD, etc.

Reviewed By: bolinfest

Differential Revision: D41321638

fbshipit-source-id: b1d71ccf2773a7806e7fe1b62296e8aac918da43