fsck: remove overlay for directories that aren't on disk
Summary:
During startup, EdenFS loads inodes for all the directories present in the
overlay, with the assumption that the state of the overlay represents the state
of the filesystem. Today, this is mostly true as the working copy only grows
and thus the only way for the working copy and overlay to shrink is to update
past commits that delete a directory.

In the future however, EdenFS will gain the ability to shrink its working copy,
either manually, or automatically. When this happens, the overlay will still
contain data for directories that are no longer in the working copy, thus
invalidating the assumption that working copy and overlay stay in sync.

To fix this, we simply need to teach the fsck code to remove overlay
directories when a directory isn't present on disk.

Reviewed By: kmancini

Differential Revision: D41007167

fbshipit-source-id: 2a6d93417995b9bb8f96c757d47b444c17ab82f8