Include `top` in `eden rage`
Summary:
There have been support requests where we needed to rule out high system load
on the user's computer (rather than EdenFS) slowing down their experience.
This adds a snapshot of `top` to the output of `eden rage` on Linux and macOS
so we don't have to ask the user for this.

We only include the top ten-ish processes in the output to avoid blowing up the
`rage`.

Reviewed By: chadaustin

Differential Revision: D41351716

fbshipit-source-id: 5db170b1b83ff58ffce21769b379e9ad2b38738a