Clean up servers when all repos have been disposed
Summary:
We've seen cases on windows where people get stale cwds due to re-using repos. We actually need to fix that problem separately, but I think it's also a good idea to reduce how aggressively we re-use servers.

If someone is running `sl isl` on the CLI, we'll pop open a new browser window regardless. It doesn't impact startup time much if we DON'T re-use the server. The only potential impact of not re-using a server is that other servers already running would get killed.

So the idea here is that we use our refcounted repository to know when we're in the clear to shutdown the entire server. Then the next `sl isl` will get a fresh server no problem.

This has the benefit of making it more likely that people will not get stuck on stale versions of ISL if we update it (though we should add other mitigations for this, too, like an in-app warning if we see that the sl version has changed since spawning)

Miscellany:
- if you spawn with --foreground, we don't do this auto cleanup, since presumably your terminal should fully control this
- We need to  consider --no-open. The idea here is that we only try to clean up after a repo is disposed. So if spawned with --no-open, it won't try to clean itself up until at least one ws connection is started succcessfully.
- We don't have a great way to write to the log file, due to some ergonomics of how we create a file logger. Ideally we'd print a little "Shutting down ISL server due to no remaining repositories + inactivity"

NOTE: This change isn't strictly necessary, but I think it's better for us to err on the side of not re-using servers. Any thoughts on this?

Reviewed By: bolinfest

Differential Revision: D41357143

fbshipit-source-id: 36f88d4a91bc6c06d511dd305b0c02cf377fd768