thrift: tweak the configuration of the server a bit
Summary:
In Thrift servers, the IO workers are merely responsible for reading from the wire and serialization and deserialization of packets. It is not much use for us to tweak that. Instead, CPU worker is responsible to pull / execute our Thrift method implementation and it's more useful to be able to change that.

One thing to note that this change does not have any material effect at all. The number of IO and CPU workers are both `std::thread::hardware_concurrency` by default in Thrift already, and that's what we have for `thrift:num-workers` in `EdenConfig`.

For `setCPUWorkerThreadName`, it will make debugging easier.

Reviewed By: MichaelCuevas

Differential Revision: D41363348

fbshipit-source-id: 0bf7af84489ce7131a1a6ed7805e996644f63f76