git: fix branches to not be public by default
Summary:
I found that after pushing a stack of commits to a new Git branch, my smartlog only showed the tip of my branch. This is because the branch became a remote name which makes its commits public (and hence not interesting).

Add a new remotenames knob "publicheads" which defines which fully qualified remote names are "public". For Git repos, it defaults to [remote/master, remote/main].

Fixes #178

Reviewed By: quark-zju

Differential Revision: D41347501

fbshipit-source-id: fe2d1e62f82a86bccbad1bdf18fad4cd5845c777