Update snapshot GC to take labels into account
Summary:
Stack context: https://fb.quip.com/FO2RAtu5ZHb7

Currently the logic for fetching expired bubbles looks at the status and the expiry date and based on those parameters, decides if the bubble is elligible for garbage collection. This diff modifies that logic to consider the presence of labels as well, i.e. if labels are present the bubble cannot be considered for cleanup. Additionally, even if the bubble has EXPIRED status or is past its expiry date, if it has labels associated with it then it should continue to be accessible. The follow-up diffs will add capability of adding and removing labels from a bubble.

Reviewed By: yancouto

Differential Revision: D40978313

fbshipit-source-id: e3b8a8d0eb33c0ba15430987a5cea67723f31b34