Include labels in snapshot show output
Summary:
Stack context: https://fb.quip.com/FO2RAtu5ZHb7

This diff modifies the response of `FetchSnapshot` to include the labels that are part of the snapshot. This change can be utilized in the `hg snapshot show` command to display the label information associated with a snapshot. Subsequent diffs will make use of this change.

Differential Revision: D41118544

fbshipit-source-id: 361a2a9dec5daea8605a0b6b7942aa65b1f4ce7a