Allow specifying the RE CAS use-case when setting the objectId
Summary:
We have a concept in Hargow collect Volumettes, which are directories of data. The backend storage for them is the RE CAS, and we would like to be able to mount volumettes using eden.

However, we have our own use-cases, which have their own storage (Manifold buckets and ZippDBs). So to make this work, we need some way to be able to specify the use-case.

We pass the use-case in using a new `--re-use-case` parameter, and store it in a `CheckoutConfig` option under `recas/use-case`.

Reviewed By: miaoyipu

Differential Revision: D41001049

fbshipit-source-id: a4b3d5c602e29a71560d17d3d3802b4b0a5922c0