Lazy loading of labels while opening or accessing snapshot
Summary:
Stack context: https://fb.quip.com/FO2RAtu5ZHb7

`Labels` in snapshots are relevant when determining their lifetime or when trying to view which labels are associated with a snapshot. However, a large number of cases do not need to rely on the labels information for their intended functionality. Additionally, these labels can be changed frequently and hence cannot be cached. This diff adds lazy loading + one time initialization for labels so that we fetch them only when needed and don't incur unnecessary cost.

Reviewed By: yancouto

Differential Revision: D41342859

fbshipit-source-id: a0ebd75a33f73ca1f0190ad6e6b3d000514f60f3