inodes: improve conflict types when invalidation fails
Summary:
Mercurial handles conflicts differently. In particular, it assumes that a
MODIFIED_MODIFIED conflict has the file present in both the source commit and
the destination commit. In some cases, we know that the source or destination
commit doesn't contain the file that is conflicting, thus let's slightly adjust
the conflict type.

Reviewed By: chadaustin

Differential Revision: D40536744

fbshipit-source-id: 58fe953fbbd92b8f15b137603e55f5651255db67