Render repo info from CodeReviewProvider
Summary:
In the next diff, we show the current repo's information in a tooltip. We want this to include the code review provider's description (repo name/owner). This requires a per-provider implementation.

Here I opt to pass the CodeReviewSystem in to each UICodeReviewProvider's contructor, so there's no need to pass the CodeReviewSystem when rendering the RepoInfo.

Reviewed By: bolinfest

Differential Revision: D41362400

fbshipit-source-id: 75775629eb3f86be37b16a7b8c01ae08897549db