Add dropdown for cwd info
Summary:
I saw one bug reported where on windows the cwd is getting confused / messed up.
It would be nice to give users more insight into their repo's state.
We don't really have this in the old ISL, but I think it's useful.

Also, in vscode we'll need a picker to let you select which repo / cwd to use based on your workspace folders (you may have multiple repos mounted as workspace folders). This will be an ideal place to put this.

I was also thinking of a button for browser-based ISL to pick a new repo, but I don't know how that would work if you tunnel ISL from a remote server...

Reviewed By: bolinfest

Differential Revision: D41362401

fbshipit-source-id: f2b5bf426219e9b4ed40e1369db24eb991ad44fc