Fix tooltips squishing and misaligning themselves on the right edge
Summary:
While working on the submit button tooltip, I ran into an issue where the tooltip was getting weirdly offset and squished that made it hard to read correctly.

The problem here has to do with how we position our tooltips.
We compute the tooltip position using top + left coordinates, then we use `position: absolute` + `left:` + `right:`. We have some code that looks if this will overflow the position off the screen, then re-adjusts to add an additional offset inwards.

The problem is if we render a tooltip very close to the right edge of the screen (like with the submit button).
If the window width is 1000px, we position it with (for example) `left: 990px`. Then, it renders itself to not overflow and ends up shrinking to only use 10px of width, even though the tooltip contents would be more comfortable with larger width.

Instead, if we were to use `right:` to position the tooltip, it would render with `right: 10px`, then it would be able to expand to its preferred width leftwards, which looks much more natural and doesn't squish it.

Of course, just switching to `right` would lead to the same problem on the left edge. So we look at the window width, and use `right` if the tooltip is on the right half of the screen.

Technically this problem also affects `top` positioning, but it's only really noticeable if the tooltip positioning were `left` or `right`, which we rarely do in similar situations. We can put the mitigation in anyway.

Note that we still compute everything along the way with `left` and `top` for simplicity. It's only at the final positioning style insertion that we need to swap to `right`. This is a simple perhaps hacky fix that works quite well in all the example situations I can find.

Reviewed By: bolinfest

Differential Revision: D41390288

fbshipit-source-id: a47e32cec8992666019319d2d378426e6fac9c36