Send cwd in repoInfo, use in client
Summary:
I noticed that the cwd was not rendering in the cwd picker in vscode. This is because we were using initialParams.get('cwd'), which is not accurate for vscode. This is also going to be wrong in the near future when we allow changing the cwd in the UI for vscode.

So instead, let's send the cwd from the server. This feel natural to me to be included next to RepoInfo (but not inside of it, because RepoInfo is re-used across cwds for the same repo).

To do this, I moved the state we fetch from the server to a separate atom, then made repositoryInfo and servercwd selectors from that atom.

Reviewed By: bolinfest

Differential Revision: D41392857

fbshipit-source-id: 6b243d61f36c5ceda5cb22ee45c518eb4a701a25