Disable submit button for unknown code review providers
Summary:
We allow repos which do not use github, but we don't actually support submitting for code review with them just yet.

In the near future, we should probably allow some sort of of `sl push` command to "submit" PRs.

Until then, we should disable the submit button, and add an explanation tooltip. We can do this for both unknown code review providers and cases with NO confgured code review provider.

Reviewed By: bolinfest

Differential Revision: D41390434

fbshipit-source-id: 6db8954a93c0f0f3e0cedc816fbefebbe5597e83