backingstore: look up blobs by IDs and ignore names
Summary:
For object lookup, names are vestigial at this point. For now, don't
bother passing them down through. It's possible that at some point in
the future we'll want names again, so my plan is to leave the Request
abstraction in place, and standardize its conversion from HgProxyHash,
which will likely be renamed as HgObjectId.

Reviewed By: genevievehelsel

Differential Revision: D41033644

fbshipit-source-id: 35555502d0989c10a79ad623997e79e233c1510f