use peaceiris/actions-gh-pages@v3 to publish reviewstack.dev
Summary:
To date, I have been deploying https://reviewstack.dev manually,
so it is finally time to use a GitHub Action instead.
For now, it is configured with `workflow_dispatch`, so we do
have to push a button to trigger a build. We can consider a weekly
release cadence or somesuch going forward.

Reviewed By: evangrayk

Differential Revision: D41393330

fbshipit-source-id: 61aeb99299a9c3bf25f94b02bf7d164d021c73a4