Show spinner while comparison view is fetching
Summary:
In the comparison view, we were showing a spinner on first fetch. But if you closed & re-opened the comparison, it wouldn't show a spinner again, even though it WOULD fetch again. This could cause the comparison view to "pop" with the newly fetched contents.

I think it's a much nicer experience to show a spinner immediately while it's fetching, which makes the pop less unexpected IMO.

Another option would be to hide the previous comparison view contents when closing and re-opening. However, I like that it caches the data and keeps it around. It makes the experience feel much faster.

I also added a refresh button, to manually trigger a refetch. And in the tooltip I note that we don't auto-refresh this view, to hint that you need to click refresh if you want a fresh comparison. (Why don't we auto-refresh? I think it would be too noisy and unreliable for whole stack comparisons, plus maybe you wanted to keep a stable view of the diff while you edit other stuff?)

I find recoil is not great at this kind of data fetching, since it seemingly assumes you use suspense to render loading state, which prevents us from showing the last cached contents. I found a sneaky way by adding a custom hook and using `set` on the comparison data to trigger refetches. This is nicer than having to trigger fetches

Reviewed By: bolinfest

Differential Revision: D41394104

fbshipit-source-id: 36afc36f3fef55e9a9d780e0d6e8968f692dbebf