doctor: invalidate all inodes on high inode count
Summary:
One of the top issues that engineers are faced with are slow update/rebase
alongside with slow startup. Both of which are a consequence of having tons of
files populated in the repo, which often comes from tools crawling the repo. As
a mitigation, EdenFS recently gained the ability to invalidate all
non-materialized files/directories, let's start by using this in `eden doctor`
when we consider that the user has a large amount (over 1M) of inodes in their
repo.

A future step will run this invalidation as a periodic background task taking
into consideration how frequent a file was accessed.

Reviewed By: kmancini

Differential Revision: D41010836

fbshipit-source-id: 14f4c5322f3879d830048ba924d5769631809bb1