Mitigate drawer styling bug when modal is visible
Summary:
Super weird issue, I think it might be a browser bug. I get slightly different behavior on chrome vs safari... Very hard to pin down exactly what's going wrong.

The bug: When you open the comparison view, the right hand drawer with commit info adds a 5px blank gap on the right side, while staying the same width. This is visible underneath the comparison view, and contributes to opening the modal making the UI look somewhat jumpy when opening a comparison.

I think this is something about display: flex or CSS variables. Weirdly, this reproduces when setting `display: none` on the modal. So the mere presence of the modal div causes other parts of the UI to change. I don't *think* there's any change to css variable values... they're also not in the same subtree of the DOM so I don't see how they would affect each other...

Anyway, my "fix" here is just to `justify-content: flex-end`, which moves the extra 5px of whitespace to the other side of the drawer, where it is less noticeable. Not a real fix by any means, but it's at least slightly less jumpy.

Reviewed By: bolinfest

Differential Revision: D41416942

fbshipit-source-id: c25b1878eecd5a13bcb9fc51a6e88d68720bb28e