add GitHub Enterprise support to the CLI commands
Summary:
This updates Sapling so that `sl ghstack` and `sl pr` work with
GitHub Enterprise repositories.

The first important change was in `github_repo_util.py` where

```
def is_github_repo(repo) -> bool:
```

was replaced with:

```
def check_github_repo(repo) -> GitHubRepo:
```

Where `check_github_repo()` looks at the URI, extracts the
hostname, and calls the following to check whether the repository
is part of a GitHub Enterprise account:

```
gh auth status --hostname <hostname>
```

The next step was to find all of the places where we have some sort
of `PullRequest` object and ensure the `hostname` is passed along
in addition to the `owner` and `name` fields. Note that if the user has
existing data written to the `PullRequestStore`, the `hostname` property
may not have been written, so we must fill in `"github.com"` in that case.

I also moved some of the functions in `github_repo_util.py` to a new file, `pr_parser.py`,
because `edenscm/ext/ghstack/__init__.py` needs `github_repo_util.py` for
`check_github_repo()`, so I wanted to reduce the number of transitive
dependencies that pulls in.

Reviewed By: muirdm

Differential Revision: D41407596

fbshipit-source-id: d66321fc9891520b08b9c03219f4f23193f15e6a