Don't support buck1 on incremental runner and fix buck2 path
Summary:
The setup script already doesn't support buck1, so no reason for this script to do so.

For the path, I've noticed this changed and doesn't work without copying the path, so I fixed it. I know we shouldn't really depend on the output format, but this script is only used by devs and not automation, so it's fine to be a little hacky.

Reviewed By: RajivTS

Differential Revision: D41400087

fbshipit-source-id: d8961eec40a394da40bbf64bf7582160940e9dd5