Address Mononoke server crashes on shutdown
Summary:
Shutdown for `Mononoke` server is based on signalling a oneshot and setting an atomic bool to true. This is then used to wind down the processes that are holding references to `MononokeApp` and `MononokeRepos` which contain materialized `Repo` facet containers. Linked to those containers are detached processes that run in the background, e.g. `warm_bookmark_cache_updater`, `segmented_changelog`, `skiplist updater`, etc.

`Mononoke` uses an implicit shutdown model wherein dropping of all `MononokeApp` and `MononokeRepos` references results in the materialized facets terminating these background processes through their `Drop` implementations. The crash during shutdown occurs in two scenarios.

### Case - I
If `Mononoke` server is still in startup mode (i.e. microwave cache warmup is not yet completed), then even if the shutdown signal is sent it's not received since the receiver is setup after the cache warming is completed. In this case, parts of the service start shutting down but the microwave and bookmark cache warmer continue to make SQL and Manifold calls. Eventually these dependent libraries perform cleanup and any subsequent calls to MySQL or Manifold through these libraries results in a crash.

### Case - II
If `Mononoke` server has completed it's startup and has begun serving requests, then that means the shutdown signal receiver is also setup. Singalling shutdown at this point ensures that the `connection_acceptor` holding references to `MononokeApp` and `MononokeRepos` drops these references. However, in addition to the `connection_acceptor` the `ShardedProcessExecutor` also holds a reference to these structs and runs in detached mode. The shutdown signal is not propogated to the executor and hence the queries to MySQL and Manifold via Repo entities continue even after signalling shutdown leading to a similar crash.

This diff addresses the second case by ensuring that the shutdown signal is propogated to the `ShardedProcessExecutor` as well. I will create a follow-up diff to address the first case.

Reviewed By: mitrandir77

Differential Revision: D41433429

fbshipit-source-id: 94e7393ba30efdbe6c7a598830ce4c6c5d4b852d