hardcode github.com for STACK_HEADER (#222)
Summary:
It does not seem likely that `ezyang/ghstack` will exist at `https://{github_url}/ezyang/ghstack)`. I assume that the hyperlink is meant to give credit/reference to ezyang's project, so this fixes the hyperlink so that it always works even if we've clone a repo from a GHE instance.

Pull Request resolved: https://github.com/facebook/sapling/pull/222

Reviewed By: evangrayk

Differential Revision: D41442576

Pulled By: bolinfest

fbshipit-source-id: 271cc23b2d2678d9d85912dc07367ae74fd99aee