Add test that shows broken behaviour on empty tag
Summary:
This reproduces an actual issue I have when running `gitimport` on `whatsapp/server`:
`gitimport` trips-up on trying to parse an empty git tag and fails without much in terms of helpful output.
That is before being able to display the list of refs and the bonsai changesets they should map to.

Reviewed By: mitrandir77

Differential Revision: D41401956

fbshipit-source-id: 68b37f95a0a7b913c1d47b5d2260fe8ded622bce