Add context to errors
Summary:
Before this change, the executable would silently fail when trying to display
the ref mapping.
Adding this context allows the error message to be more helpful and informative.
It helped me figure out that an empty tag was the source of this particular problem, and could come in handy again next time something is broken around there.

Reviewed By: mitrandir77

Differential Revision: D41401955

fbshipit-source-id: 201f2254e371e4f6130376583d22116ad89f1951