Update ODS counter in both pass and fail scenarios
Summary: The ODS metric for tracking errors in tunables won't exist until we actually encounter errors. This diff pushes `0` value to the counter if things work as expected and `1` value to the counter if an error was encountered. This would ensure that the counter exists in all cases and we can setup alerts based on it.

Reviewed By: mitrandir77

Differential Revision: D41435634

fbshipit-source-id: bd117c982cf20ceb745ff1675a6da2f492994de1