Fixup gitoxide bug with empty tag
Summary:
Vendor a patch to `git-object` that fixes encoding of an empty git commit tag.

I [reported](https://github.com/Byron/gitoxide/issues/603) and [fixed](https://github.com/Byron/gitoxide/pull/604) this issue upstream, but the fix is only in the `main` branch for now and not yet in a release.

Reviewed By: mitrandir77

Differential Revision: D41401957

fbshipit-source-id: 27933dd6fe852f93824c4d3f4ca9c9a96d6d2faa