Add internals documentation
Summary:
Add documentation for the internals of ISL. This should make it easier for new folks to onboard to ISL or contribute from OSS.

I opted to add this to the `isl/README` file. We could also put this on the sapling-scm.com website...? Unclear to me which is the preferred method. I suppose we have the `internals` section where we could add an `ISL` page. But the README ensures this stays close to the code. Open for opinions here.

Reviewed By: akushner

Differential Revision: D41482993

fbshipit-source-id: a5066faa1ae751feb10c7ce9a99aed7cd12d7f58