Improve the contrast of Get Started for the dark theme
Summary:
When using the dark theme the contrast of the Get Started button is not very good.
This diff sets the font color of Get Started to be black the whole time like how it is in the light theme. Currently in the dark theme the font color is white by default and turns black when hovered over.

Reviewed By: MichaelCuevas

Differential Revision: D41432365

fbshipit-source-id: 03fe09a1a151e74c9e86d2b133b863ff2a845ab9