Allow remove-labels to delete all labels without input
Summary: The expectation of the `remove_bubble_labels` method was to delete the specified input labels from the bubble and in case of no input labels provided, it should delete all labels associated with the bubble. In the later case, the current implementation of the method was not working as expected. This diff fixes that behaviour.

Reviewed By: yancouto

Differential Revision: D41520824

fbshipit-source-id: 91bf312d00d6d4822e69b4cab62428f7b0a396f9