Modify remove-labels command to allow for empty labels
Summary: The expectation of the `remove-labels` command was to delete the specified input labels from the bubble and in case of no input labels provided, it should delete all labels associated with the bubble. In the later case, the current implementation of the command doesn't work. It mandates the presence of a `labels` argument for the command to work. This diff fixes that behaviour.

Reviewed By: yancouto

Differential Revision: D41521650

fbshipit-source-id: 36d237881ba18d0680893997d55e5f73dd1a05ff