migrate RepoArgs from derive into builder syntax
Summary:
I want to create a macro generating various types of RepoArgs and for doing
that kind of sorcery it's better to deal with concrete code on the clap side
rather than with another layer of macros.

This also allows to make repoargs an enum rather than struct which I think is
pretty neat. [included in the next diff]

Reviewed By: RajivTS

Differential Revision: D41469284

fbshipit-source-id: acf9f441d476e42637c26e43e442f0394834860b