backsyncer: split out backsyncer lib and cmd
Summary:
The backsyncer library and command are in the same cargo file, so their dependencies get merged together.  This can cause problems with cyclic dependencies in the event that something that the command depends on (e.g. cmdlib) becomes dependent on the library.

Split them out to separate directories and cargo files.

Reviewed By: mitrandir77

Differential Revision: D41549923

fbshipit-source-id: a09979fc484e67d5a7d517f5940982f09000fa95