control-point: add crate for reproducing race conditions
Summary:
This adds a new control-point crate which can be used for tests to
setup and test race conditions. It works as follows:

Production code contains a `control_point("some_name")` call. Internally this
does an environment variable check for CONTROL_POINT_FILE. If the environment
variable doesn't exist, the function does nothing and returns. If the variable
is set, the function attempts to read the contents of the file the envvar points
to.

That file contains json which contains instructions for what to do when each
control point is hit. Currently these instructions include "wait", "continue",
and "panic". These instruction can optionally contain arguments, such as "panic"
allowing a "message" argument.

While the control point is being processed, it writes the processing status to
the CONTROL_POINT_FILE location plus a ".response" suffix.

During a "wait", the process loops reading the instruction file until the
instruction changes.

The test code produces the instruction file to tell the production code to wait
at certain points, then reads the response file to know when the production code
has hit the control-point. It can then instruct the production code to proceed,
by changing the instruction file.

Multiple control-points with the same name and multiple threads hitting the same
control-point are currently not supported.

Reviewed By: muirdm

Differential Revision: D41239836

fbshipit-source-id: ba7448dfb1175b0accb31f58aab89b4ea39b5ab9