Log metadata size mismatch on NFS READ
Summary:
Log to text logs and Scuba if there's an NFS READ where the size of the file
contents didn't match the size in metadata.  This way we can see whether size
mismatches are still happening much in practice, in order to know how to
prioritize further investigation.

Based on kmancini's D41239395, but adds Scuba logging too.

Reviewed By: chadaustin, kmancini

Differential Revision: D41393677

fbshipit-source-id: 968c2866d8fd353e89ad871080a57308cacdc102