util: remove usage of "imp"
Summary:
"imp" is deprecated and will be removed in 3.12.

I don't think this frozen check is relevant anymore, and there doesn't seem to be an analog in importlib, so just kill it. Nowadays the sl entrypoint is Rust, so I don't think we can be in a frozen Python executable.

Reviewed By: quark-zju

Differential Revision: D41394322

fbshipit-source-id: eaeacad44d4e0fd1de08b26db7fdc431244009ed