Add commit sync context to backsync_latest
Summary:
When using the `backsync_latest` function, it used to always use `CommitSyncContext::Backsyncer`, even though that function was used from different places (like pushrebasing and scs endpoints).

This adds the option to customise it, and correctly does so in these places that are not the backsyncer.

I also reused the helper `backsync_latest` in two more places where it was possible to simplify code.

Reviewed By: mitrandir77

Differential Revision: D41401071

fbshipit-source-id: 562e8f6968a45bed331edaa27a48fd947594824d