Log pushrebase on small repo in wireproto path
Summary:
This fixes wireproto code so that it logs small commits. Using the test added on the previous diff, we can verify we now correctly log the commits.

We still don't log the bookmark updates, as that works a little differently, but should be possible.

The wireproto code is confusing to touch, a lot of cases for things that (possibly?) don't even happen anymore, I wish we can move to mononoke_api/EdenAPI soon.

Reviewed By: mitrandir77

Differential Revision: D41439445

fbshipit-source-id: 3cb316fb8cad6eaa867534abfc261bcc25611b67