SomeMissingItem.value can never be None
Summary:
When I was looking at this code, I noticed that `SomeMissingItem`, which means that a value was present in some blobstore but missing in others, can never have a `None` value, since it is present in at least a blobstore.

This diff changes the type so it's not optional, which should make things clearer.

To avoid `unwrap`s or unnecessary never-ocurring errors, I used `Option::get_or_insert_with` to store together the value and the map of values to blobstores.

Reviewed By: mitrandir77

Differential Revision: D41469264

fbshipit-source-id: 9613cc179def28d421469bcd6d89f2bbd58c4136