Warn when unlocking repo with disabled pushredirection
Summary:
This is the same behaviour that `scmadmin repo unlock` has. It is meant to be a warning to make us careful. If we unlock the repos and pushes are allowed from both sides, the repos may diverge.

I tried to make the wording a little clearer than it was previously.

This diff adds some logic to fetch current sync config from configerator, and find what "group" your repo belongs to, which will be helpful later on to lock the whole group at once.

Reviewed By: mitrandir77

Differential Revision: D41473614

fbshipit-source-id: a4047de5f740a2d32c4ee9a1a321631e96193886