Add vscode config for the command to use for `sl`
Summary:
When spawning ISL from the CLI, we allow `--command` / infer the command executable from the `sl` in `sl web`.

In vscode, we don't have the same direct place to infer the command from. `sl` is a good guess, and works if you've installed `sl` on your path. But we should allow a config to control exactly what installation / path we spawn ISL from.

This may help with #282

Reviewed By: bolinfest

Differential Revision: D41590875

fbshipit-source-id: 8b909e527e5f0152396ef9263edbeff3eca11f9b