Detect ENOENT on windows
Summary:
I thought we could depend on execa to manage returning `ENOENT` for invalid command names, but apprently this doesn't always happen on windows.

By making this test work on windows, we show a more accurate "sl is not installed correctly" rather than "this is an invalid repository".

See this discussion for more information:
https://github.com/sindresorhus/execa/issues/469#issuecomment-859924543

Example of a user on windows not triggering ENOENT:
https://github.com/facebook/sapling/issues/282#issuecomment-1331315208

Instead, we should detect invalid commands on windows with exit code 1 (but only on windows).

The use of this detection is isolated to findRoot, which is where we determine if the command being used is valid. If some other issue triggered this code path, showing "this is not a valid sl command" is also a reasonable error message. Plus we still log the error for debugging purposes.

This may also help with #282.

Reviewed By: bolinfest

Differential Revision: D41595211

fbshipit-source-id: 7cff8eafd58f6da14bb85f4289eb25ddbcf5273c