refactor --all-sources flag to be --only-repo-source
Summary:
The --all-sources flag doesn't really make sense, and it hasn't been used internally within the past 90 days...

I'm open to killing this flag altogether, but for now we can just change the logic so that it makes sense.

Now we have renamed the flag to --only-repo-source (more accurate), default this flag to false, and require users to pass it in if they want to change the behavior of redirect fixup.

Reviewed By: fanzeyi

Differential Revision: D41488196

fbshipit-source-id: 5f36c2199e9eb849cdc3db6ce163462b3e295b47